Moved and fixed errors
