+++ frontend/app/Domits/src/screens/message/components/chatMessage.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { View, Text, Image, StyleSheet } from 'react-native';
+import { styles } from '../styles/chatMessageStyles';
 
 const ChatMessage = ({ message, userId, contactName, }) => {
     const { userId: senderId, text, createdAt, fileUrls = [] } = message;
@@ -19,11 +20,11 @@ const ChatMessage = ({ message, userId, contactName, }) => {
             ]}
         >
             <View style={styles.messageContent}>
-                <View style={isCurrentUser ? styles.youHeader : styles.senderHeader}>
+                <View style={isCurrentUser ? styles.userHeader : styles.senderHeader}>
                     {!isCurrentUser && (
                         <>
                             <Image
-                                source={require('../pictures/domits-logo.jpg')}
+                                source={require('../../pictures/domits-logo.jpg')}
                                 style={styles.senderImage}
                             />
                             <Text style={styles.senderName}>{contactName}</Text>
@@ -33,17 +34,17 @@ const ChatMessage = ({ message, userId, contactName, }) => {
 
                     {isCurrentUser && (
                         <>
-                            <Text style={styles.youMessageDate}>{formatDate(createdAt)}</Text>
-                            <Text style={styles.youLabel}>YOU</Text>
+                            <Text style={styles.userMessageDate}>{formatDate(createdAt)}</Text>
+                            <Text style={styles.userLabel}>YOU</Text>
                             <Image
-                                source={require('../pictures/domits-logo.jpg')}
-                                style={styles.youImage}
+                                source={require('../../pictures/domits-logo.jpg')}
+                                style={styles.userImage}
                             />
                         </>
                     )}
                 </View>
 
-                <Text style={isCurrentUser ? styles.youMessageText : styles.senderMessageText}>
+                <Text style={isCurrentUser ? styles.userMessageText : styles.senderMessageText}>
                     {text}
                 </Text>
 
@@ -80,94 +81,6 @@ const ChatMessage = ({ message, userId, contactName, }) => {
     );
 };
 
-const styles = StyleSheet.create({
-    messageContainer: {
-        flexDirection: 'row',
-        alignItems: 'flex-end',
-        marginVertical: 5,
-    },
-    messageLeft: {
-        alignSelf: 'flex-start',
-    },
-    messageRight: {
-        alignSelf: 'flex-end',
-        flexDirection: 'row-reverse',
-    },
-    messageContent: {
-        maxWidth: '95%',
-        padding: 10,
-        borderRadius: 10,
-        backgroundColor: 'white',
-        width: '100%',
-    },
-    senderHeader: {
-        flexDirection: 'row',
-        marginBottom: 5,
-        alignItems: 'center',
-    },
-    youHeader: {
-        flexDirection: 'row',
-        marginBottom: 5,
-        alignItems: 'center',
-        alignSelf: 'flex-end',
-    },
-    senderImage: {
-        width: 40,
-        height: 40,
-        borderRadius: 20,
-        marginRight: 5,
-    },
-    youImage: {
-        width: 40,
-        height: 40,
-        borderRadius: 20,
-        marginRight: 5,
-    },
-    senderName: {
-        fontSize: 20,
-        color: 'black',
-        fontWeight: '500',
-        marginLeft: 10,
-    },
-    youLabel: {
-        fontSize: 16,
-        color: 'black',
-        fontWeight: '500',
-        marginRight: 10,
-    },
-    senderMessageText: {
-        fontSize: 16,
-        color: '#333',
-        marginLeft: 16,
-        marginTop: 2,
-        fontWeight: '400',
-        borderColor: 'green',
-        paddingTop: 8,
-        maxWidth: '90%',
-    },
-    youMessageText: {
-        fontSize: 16,
-        color: '#333',
-        marginRight: 15,
-        marginTop: 2,
-        alignSelf: 'flex-end',
-        paddingTop: 8,
-        maxWidth: '100%',
-    },
-    senderMessageDate: {
-        fontSize: 12,
-        color: 'black',
-        marginLeft: 10,
-        alignSelf: 'center',
-        fontWeight: '500',
-    },
-    youMessageDate: {
-        fontSize: 12,
-        color: 'black',
-        alignSelf: 'center',
-        marginRight: 10,
-        fontWeight: '500',
-    },
-});
+
 
 export default ChatMessage;
