+++ frontend/app/Domits/src/screens/message/components/contactItem.js
@@ -1,7 +1,8 @@
-const profileImage = require('../pictures/domits-logo.jpg');
-import React from 'react';
-import { View, Text, Image, StyleSheet } from 'react-native';
+const profileImage = require('../../pictures/domits-logo.jpg');
 
+import React from 'react';
+import { View, Text, Image, } from 'react-native';
+import { styles } from '../styles/contactItemStyles';
 
 const ContactItem = ({ contact, isPending, selected, dashboardType }) => {
     const isGuest = dashboardType === 'guest';
@@ -28,40 +29,5 @@ const ContactItem = ({ contact, isPending, selected, dashboardType }) => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        marginTop: 0,
-        paddingVertical: 20,
-        paddingHorizontal: 10,
-        borderBottomWidth: 0.5,
-        borderColor: '#e0e0e0',
-    },
-    
-   
-    profileImage: {
-        width: 35,
-        height: 35,
-        borderRadius: 25,
-        marginRight: 15,
-        backgroundColor: '#e0e0e0',
-    },
-    textContainer: {
-        flex: 1,
-    },
-    fullName: {
-        fontSize: 18,
-        fontWeight: 'bold',
-        color: '#333',
-    },
-    subtitle: {
-        fontSize: 15,
-        color: '#555',
-        marginTop: 3,
-
-    },
-});
 
 export default ContactItem;
